requirements.txt
Package                       Version
----------------------------- -------
asgiref                       3.5.2
atomicwrites                  1.4.1
attrs                         21.4.0
black                         22.6.0
click                         8.1.3
colorama                      0.4.5
coverage                      6.4.1
Django                        4.0.6
django-extensions             3.1.5
django-filter                 22.1
djangorestframework           3.13.1
djangorestframework-simplejwt 5.2.0
drf-nested-routers            0.93.4
flake8                        4.0.1
flake8-django                 1.1.1
flake8-html                   0.4.2
iniconfig                     1.1.1
Jinja2                        3.1.2
MarkupSafe                    2.1.1
mccabe                        0.6.1
mypy-extensions               0.4.3
packaging                     21.3
pathspec                      0.9.0
pip                           22.1.2
pip-chill                     1.0.1
platformdirs                  2.5.2
pluggy                        1.0.0
psycopg2                      2.9.3
py                            1.11.0
pycodestyle                   2.8.0
pyflakes                      2.4.0
Pygments                      2.12.0
PyJWT                         2.4.0
pyparsing                     3.0.9
pytest                        7.1.2
pytest-cov                    3.0.0
pytest-django                 4.5.2
python-decouple               3.6
pytz                          2022.1
setuptools                    58.1.0
sqlparse                      0.4.2
tomli                         2.0.1
tzdata                        2022.1
